shared/ssa.c: Trivial patch to eliminate compile warning
src/ssa.c: In function #ssa_start_svc#:
src/ssa.c:662: warning: format #%llu# expects type #long long unsigned int#, but argument 5 has type #uint64_t#

+++ shared/ssa.c
@@ -659,7 +659,8 @@ struct ssa_svc *ssa_start_svc(struct ssa_port *port, uint64_t database_id,
 
 	svc->index = port->svc_cnt;
 	svc->port = port;
-	snprintf(svc->name, sizeof svc->name, "%s:%llu", port->name, database_id);
+	snprintf(svc->name, sizeof svc->name, "%s:%llu", port->name,
+		 (unsigned long long) database_id);
 	svc->database_id = database_id;
 	svc->rsock = -1;
 	svc->state = SSA_STATE_IDLE;
